had to adapt to the breaking change in ceres 1.7.0 :
options->sparse_linear_algebra_library
is since then 
options->sparse_linear_algebra_library_type

Of course I tried to use the CERES_VERSION macro to make the code compile with pre and past 1.7.0 ceres version. But some moron decided to define 
the version macro to be just 1.7.0 !!!!!! There is just _no_ way to use
this in a preprocessor #if! Can somebody believe this? C coders know for ages to use integral literal as version macros to enable the preprocessor to do just the right thing with it. But 1.7.0 is far from a integral literal. But google's ceres development team does not know about it??? Is this a bad joke?